Continue CSV transformer

+++ src/eds-transforms/src/main/java/org/endeavourhealth/transform/emis/csv/transforms/coding/ClinicalCodeTransformer.java
@@ -40,7 +40,7 @@ public abstract class ClinicalCodeTransformer {
 
         ClinicalCode c = new ClinicalCode(emisTerm, emisCode, emisCategory,
                 nationalCategory, nationalCode, nationalDescription,
-                snomedConceptId.longValue(), snomedDescriptionId.longValue());
+                snomedConceptId, snomedDescriptionId);
 
         map.put(codeId, c);
     }
